requirements.txt
django==1.8.7
pillow
unidecode
django-allauth
django-ckeditor
django-formset-js
django-haystack
django-jquery-js
django-mptt
django-suit
django_mptt_admin
elasticsearch
pip
setuptools
future
virtualenv
simplejson
django-daterange-filter
django-import-export
django-scheduler
django-cms
django-treebeard
django-sekizai
django-classy-tags
djangocms-admin-style
html5lib
six
django-filer
cmsplugin-filer
django-reversion
djangocms_file
djangocms_flash
djangocms_googlemap
djangocms_inherit
djangocms_picture
djangocms_teaser
djangocms_video
djangocms_link
djangocms_snippet
djangocms_text_ckeditor
django-report-builder


